refactor: remove deprecated files and update MultiUpstreamPort implementation
- Delete outdated documentation and test files related to old interfaces.
- Update MultiUpstreamPort to ensure it only accepts a shared channel for upstream ports.
- Ensure all methods raise appropriate panics for invalid usage.
- Simplify the architecture documentation to reflect the current design.

+++ internal/core/cycle_port/cycle_port_test.go
@@ -1152,5 +1152,5 @@ func TestUpstreamCycleIncrementMatchesNonReadyCount(t *testing.T) {
 	}
 
 	t.Logf("Cycle increment matches non-ready count: original %d, final %d, incremented by %d",
-		originalCycle, receivedPkt.Cycle, receivedPkt.Cycle-uint64(originalCycle))
+		originalCycle, receivedPkt.Cycle, receivedPkt.Cycle-originalCycle)
 }
